update examples to new I/O format. All tests passing except new DataFile tests

+++ datafs/core/data_file.py
@@ -4,8 +4,7 @@ import fs.path
 import tempfile
 import shutil
 import time
-# from fs.osfs import OSFS
-from fs.tempfs import TempFS
+from fs.osfs import OSFS
 from fs.multifs import MultiFS
 
 from fs.errors import (ResourceLockedError, ResourceInvalidError)
@@ -36,7 +35,7 @@ def clear_outdated_cache_files(cache, service_path, latest_version_check):
 
     # Check to see if there is a valid system file in the cache at the service 
     # path. If so, check if it is up to date. If not, raise NoCacheError.
-    if cache.fs.exists(service_path) and cache.fs.hassyspath():
+    if cache.fs.exists(service_path) and cache.fs.hassyspath(service_path):
         up_to_date = latest_version_check(cache.fs.getsyspath(service_path))
 
     else:
@@ -44,15 +43,23 @@ def clear_outdated_cache_files(cache, service_path, latest_version_check):
     
     # Delete the file in the local cache if it is out of date.
     if not up_to_date:
-        self.cache.fs.remove(self.service_path)
+        cache.fs.remove(service_path)
 
 def determine_cache_state(authority, cache, service_path, latest_version_check):
 
     try:
         clear_outdated_cache_files(cache, service_path, latest_version_check)
-        if not cache.isfile(service_path):
+        
+        if (not cache.fs.isfile(service_path)) and authority.fs.isfile(service_path):
+            cache.fs.makedir(
+                fs.path.dirname(service_path),
+                recursive=True,
+                allow_recreate=True)
+
             fs.utils.copyfile(authority.fs, service_path, cache.fs, service_path)
+
         use_cache = True
+
     except NoCacheError:
         use_cache = False
 
@@ -69,30 +76,21 @@ def create_downloader(authority, cache, use_cache):
 
     return fs_wrapper
 
-def create_writer(service_path):
-    
-    write_fs = TempFS()
-    write_fs.makedir(
-        fs.path.dirname(service_path),
-        recursive=True,
-        allow_recreate=True)
-
-    return write_fs
 
 def close(filesys):
 
     closed = False
 
-        for i in range(5):
-            try:
-                filesys.close()
-                closed = True
-                break
-            except ResourceLockedError as e:
-                time.sleep(0.5)
+    for i in range(5):
+        try:
+            filesys.close()
+            closed = True
+            break
+        except ResourceLockedError as e:
+            time.sleep(0.5)
 
-        if not closed:
-            raise e
+    if not closed:
+        raise e
 
 
 
@@ -153,24 +151,38 @@ def open_file(authority, cache, update, service_path, latest_version_check, *arg
 
     use_cache : bool
 
-         update, service_pathh, latest_version_check, **kwargs
+         update, service_path, latest_version_check, **kwargs
     '''
 
     use_cache = determine_cache_state(authority, cache, service_path, latest_version_check)
     fs_wrapper = create_downloader(authority, cache, use_cache)
-    write_fs = create_writer(service_path)
+    
+    tmp = tempfile.mkdtemp()
+    try:
+        write_fs = OSFS(tmp)
+        write_fs.makedir(
+            fs.path.dirname(service_path),
+            recursive=True,
+            allow_recreate=True)
 
-    fs_wrapper.setwritefs(write_fs)
+        fs_wrapper.setwritefs(write_fs)
 
-    f = fs_wrapper.open(service_path, *args, **kwargs)
-    yield f
-    f.close()
 
-    if write_fs.isfile(service_path):
-        update(write_fs.getsyspath(service_path))
+        try:
+            
+            f = fs_wrapper.open(service_path, *args, **kwargs)
+            yield f
 
-    close(write_fs)
+            f.close()
 
+            if write_fs.isfile(service_path):
+                update(write_fs.getsyspath(service_path))
+
+        finally:
+            close(write_fs)
+
+    finally:
+        shutil.rmtree(tmp)
 
 @contextmanager
 def get_local_path(authority, cache, update, service_path, latest_version_check):
@@ -224,38 +236,47 @@ def get_local_path(authority, cache, update, service_path, latest_version_check)
 
     use_cache : bool
 
-         update, service_pathh, latest_version_check, **kwargs
+         update, service_path, latest_version_check, **kwargs
     '''
 
     
     use_cache = determine_cache_state(authority, cache, service_path, latest_version_check)
     fs_wrapper = create_downloader(authority, cache, use_cache)
-    write_fs = create_writer(service_path)
-
-    if use_cache and cache.fs.isfile(service_path):
-        fs.utils.movefile(cache.fs, service_path, write_fs, service_path)
     
-    elif fs_wrapper.isfile(service_path):
-        fs.utils.copyfile(fs_wrapper, service_path, write_fs, service_path)
-
-    else:
-        write_fs.createfile(service_path)
-
-    local_path = write_fs.getsyspath(service_path)
-
-    yield local_path
-
+    tmp = tempfile.mkdtemp()
     try:
-        if write_fs.isfile(service_path):
-            if not latest_version_check(write_fs.getsyspath(service_path)):
-                update(write_fs.getsyspath(service_path))
+        write_fs = OSFS(tmp)
+        write_fs.makedir(
+            fs.path.dirname(service_path),
+            recursive=True,
+            allow_recreate=True)
 
-            else:
-                raise ValueError('I got here')
+
+        if use_cache and cache.fs.isfile(service_path):
+            fs.utils.movefile(cache.fs, service_path, write_fs, service_path)
+        
+        elif fs_wrapper.isfile(service_path):
+            fs.utils.copyfile(fs_wrapper, service_path, write_fs, service_path)
 
         else:
-            raise OSError('Local file removed during execution. Archive not updated.') 
+            write_fs.createfile(service_path)
+
+        local_path = write_fs.getsyspath(service_path)
+
+        yield local_path
+
+        try:
+            if write_fs.isfile(service_path):
+                if not latest_version_check(write_fs.getsyspath(service_path)):
+                    update(write_fs.getsyspath(service_path))
+
+            else:
+                raise OSError('Local file removed during execution. Archive not updated.') 
+
+        finally:
+
+            close(write_fs)
+
 
     finally:
-
-        close(write_fs)
+        shutil.rmtree(tmp)
